Modify nsexec/include/nsexec.h
