+++ nsexec/CMakeLists.txt
@@ -1,0 +1,4 @@
+cmake_minimum_required(VERSION 3.13)
+project(nsexec C)
+add_subdirectory(src)
+add_subdirectory(test)
