+++ nsexec/src/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(nsexec STATIC nsexec.c)
+target_include_directories(nsexec PUBLIC ${PROJECT_SOURCE_DIR}/include)
